# openapi.yaml
openapi: 3.1.0
info:
  title: Agung Presence API
  version: 1.0.0
  description: API contract for the Agung Presence Laravel backend.
servers:
  - url: "{{base_url}}"
...
paths:
...
  /api/attendance-requests/{attendanceRequest}/review:
    parameters:
      - $ref: "#/components/parameters/AttendanceRequestId"
    patch:
      tags:
        - Attendance Requests
      summary: Review Request
      description: Administrator-only endpoint. Approval creates attendance rows for the request date range; rejection requires a reason.
      operationId: reviewAttendanceRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewAttendanceRequest"
            examples:
              approve:
                value:
                  approval_status: approved
              reject:
                value:
                  approval_status: rejected
                  rejection_reason: Bukti tidak sesuai.
      responses:
        "200":
          description: Attendance request reviewed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRequestResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Sanctum token
      description: Use the Postman environment variable `{{token}}` as the bearer token value.
  parameters:
...
    AttendanceRequestId:
      name: attendanceRequest
      in: path
      required: true
      description: Attendance request ID.
      schema:
        type: integer
        format: int64
  responses:
    Unauthenticated:
      description: Unauthenticated.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        data:
          nullable: true
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    ReviewAttendanceRequest:
      type: object
      required:
        - approval_status
      properties:
        approval_status:
          type: string
          enum:
            - approved
            - rejected
        rejection_reason:
          type:
            - string
            - "null"
          maxLength: 2000
          description: Required when `approval_status` is `rejected`.
    UserRole:
      type: string
      enum:
        - administrator
        - employee
...
    AttendanceRequestStatus:
      type: string
      enum:
        - pending
        - approved
        - rejected
    User:
      type: object
      required:
        - id
        - name
        - username
        - email
        - role
        - created_at
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        created_at:
          type:
            - string
            - "null"
          format: date-time
...
    AttendanceRequest:
      type: object
      required:
        - id
        - user_id
        - type
        - start_date
        - end_date
        - description
        - proof_photo
        - approval_status
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        type:
          type: string
          enum:
            - sick
            - leave
            - permit
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        description:
          type: string
        proof_photo:
          type: string
        approval_status:
          $ref: "#/components/schemas/AttendanceRequestStatus"
        reviewed_by:
          type:
            - integer
            - "null"
          format: int64
        reviewed_at:
          type:
            - string
            - "null"
          format: date-time
        rejection_reason:
          type:
            - string
            - "null"
        created_by:
          type:
            - string
            - "null"
        updated_by:
          type:
            - string
            - "null"
        created_at:
          type:
            - string
            - "null"
          format: date-time
        updated_at:
          type:
            - string
            - "null"
          format: date-time
        user:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        reviewer:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
...
    AttendanceRequestResponse:
      type: object
      required:
        - message
        - data
      properties:
        message:
          type: string
        data:
          $ref: "#/components/schemas/AttendanceRequest"
...
security:
  - sanctum: []